Fix OG image generation for blog posts
The file-based opengraph-image.tsx approach has known issues with
dynamic routes in Next.js 16. This commit switches back to the more
reliable route handler approach while maintaining the centralized
metadata in layout.tsx.

Changes:
- Restore /og/[id] route handler for reliable OG image generation
- Update layout.tsx to explicitly reference /og/[id] in metadata
- Remove problematic opengraph-image.tsx from dynamic route
- Add proper metadata exports to root opengraph-image.ts
- Document new blog post publishing workflow in CLAUDE.md and README.md

The centralized metadata approach in posts.json and layout.tsx is
preserved, eliminating the need for metadata exports in individual
MDX files.
